Refactored Epics' locations
