+++ client/store/epics/combinedEpics.js
@@ -1,7 +1,7 @@
 import { combineEpics } from "redux-observable";
 
 import { loginUser } from "./userEpics/loginEpics";
-import { getRooms } from "./userEpics/roomEpics";
-import { initConnection } from "./utilEpics/connectionEpics";
+import { getRooms } from "./roomEpics";
+import { initConnection } from "./connectionEpics";
 
 export default combineEpics(initConnection, loginUser, getRooms);
